Initialize CVS branch cherry-xenmp
The per-file CVS branch roots do not exactly match a reconstructed source-branch commit.

+++ regress/sys/arch/i386/ldt/testldt.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: testldt.c,v 1.13 2005/12/24 21:43:51 perry Exp $	*/
+/*	$NetBSD: testldt.c,v 1.15 2011/03/18 03:06:21 joerg Exp $	*/
 
 /*-
  * Copyright (c) 1993 The NetBSD Foundation, Inc.
@@ -12,13 +12,6 @@
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 3. All advertising materials mentioning features or use of this software
- *    must display the following acknowledgement:
- *	This product includes software developed by the NetBSD
- *	Foundation, Inc. and its contributors.
- * 4. Neither the name of The NetBSD Foundation nor the names of its
- *    contributors may be used to endorse or promote products derived
- *    from this software without specific prior written permission.
  *
  * THIS SOFTWARE IS PROVIDED BY THE NETBSD FOUNDATION, INC. AND CONTRIBUTORS
  * ``AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
@@ -65,7 +58,7 @@ get_cs(void)
 {
 	unsigned short _v;
 
-	__asm ("movw %%cs,%0": "=r" ((unsigned short) _v));
+	__asm ("movw %%cs,%0": "=r" (_v));
 	return _v;
 }
