+++ usr.bin/pmap/pmap.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: pmap.h,v 1.6 2008/01/02 17:23:32 yamt Exp $ */
+/*	$NetBSD: pmap.h,v 1.8 2009/04/13 00:27:38 lukem Exp $ */
 
 /*
  * Copyright (c) 2002, 2003 The NetBSD Foundation, Inc.
@@ -15,13 +15,6 @@
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 3. All advertising materials mentioning features or use of this software
- *    must display the following acknowledgement:
- *      This product includes software developed by the NetBSD
- *      Foundation, Inc. and its contributors.
- * 4. Neither the name of The NetBSD Foundation nor the names of its
- *    contributors may be used to endorse or promote products derived
- *    from this software without specific prior written permission.
  *
  * THIS SOFTWARE IS PROVIDED BY THE NETBSD FOUNDATION, INC. AND CONTRIBUTORS
  * ``AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
@@ -99,7 +92,7 @@
 
 /* suck the data from the kernel */
 #define _KDEREFOK(kd, addr, dst, sz) \
-	(kvm_read((kd), (addr), (dst), (sz)) == (sz))
+	((size_t)kvm_read((kd), (addr), (dst), (sz)) == (size_t)(sz))
 #define _KDEREF(kd, addr, dst, sz) do { \
 	if (!_KDEREFOK((kd), (addr), (dst), (sz))) \
 		errx(1, "trying to read %lu bytes from %lx: %s", \
