+++ usr.sbin/eeprom/eehandlers.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: eehandlers.c,v 1.13 2007/01/16 17:32:04 hubertf Exp $	*/
+/*	$NetBSD: eehandlers.c,v 1.15 2009/04/30 07:45:28 nakayama Exp $	*/
 
 /*-
  * Copyright (c) 1996 The NetBSD Foundation, Inc.
@@ -15,13 +15,6 @@
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 3. All advertising materials mentioning features or use of this software
- *    must display the following acknowledgement:
- *        This product includes software developed by the NetBSD
- *        Foundation, Inc. and its contributors.
- * 4. Neither the name of The NetBSD Foundation nor the names of its
- *    contributors may be used to endorse or promote products derived
- *    from this software without specific prior written permission.
  *
  * THIS SOFTWARE IS PROVIDED BY THE NETBSD FOUNDATION, INC. AND CONTRIBUTORS
  * ``AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
@@ -185,7 +178,7 @@ ee_num8(ktent, arg)
 	int i;
 
 	if (arg) {
-		for (i = 0; i < (strlen(arg) - 1); ++i)
+		for (i = 0; i < (int)strlen(arg) - 1; ++i)
 			if (!isdigit((unsigned char)arg[i]))
 				BARF(ktent);
 		num32 = atoi(arg);
@@ -211,7 +204,7 @@ ee_num16(ktent, arg)
 	int i;
 
 	if (arg) {
-		for (i = 0; i < (strlen(arg) - 1); ++i)
+		for (i = 0; i < (int)strlen(arg) - 1; ++i)
 			if (!isdigit((unsigned char)arg[i]))
 				BARF(ktent);
 		num32 = atoi(arg);
@@ -391,7 +384,7 @@ ee_kbdtype(ktent, arg)
 	int i;
 
 	if (arg) {
-		for (i = 0; i < (strlen(arg) - 1); ++i)
+		for (i = 0; i < (int)strlen(arg) - 1; ++i)
 			if (!isdigit((unsigned char)arg[i]))
 				BARF(ktent);
 		kbd2 = atoi(arg);
